Fixed layout body class
Turns out `page.layout` only surfaces layouts explicitly set in the
page frontmatter, so it’s empty if a default layout (from config.yml)
is implicitly used. To resolve this, we check for empty `page.layout`
and manually output class `layout-standard` for it.

Issues with this: (1) If the default layout changes, I need to manually
change it in this file too. (2) Can’t use more complex defaults, e.g.
different default layout values depending on the page’s scope.